requirements.txt
c3d==0.5.2
lxml==5.1.0
matplotlib==3.8.3
notebook==7.1.2
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
python-dotenv==1.0.0
scipy==1.14.1
